Module chattium_oxide_lib::json

JSON bi-directional serialization.

The FromJsonnable trait provides an abstract way of handling types, that can be deserialized from JSON

The ToJsonnable trait provides an abstract way of handling types, that can be serialized to JSON

Examples

Consider a user information struct UserInfo that needs to be stored somehow:

extern crate serde;
extern crate serde_json;
extern crate chattium_oxide_lib;

use serde::de::{Error, Type};
use serde_json::value::Value;
use serde_json::error::Error as JsonError;
use serde_json::builder::ObjectBuilder;
use chattium_oxide_lib::json::{ToJsonnable, FromJsonnable};

#[derive(Debug, PartialEq)]
struct UserInfo {
    name: String,
    id: i64,
    address: (String, String),
}

impl ToJsonnable for UserInfo {
    fn to_json(&self) -> Value {
        ObjectBuilder::new().insert("name", &self.name)
                            .insert("id", &self.id)
                            .insert("address", &self.address)  // Tuples serialize to a JSON array
                            .build()
    }
}

impl FromJsonnable for UserInfo {
    // Verbose, but safe, matching algorithm
    fn from_json(json: Value) -> Result<Self, JsonError> {
        match json {
            Value::Object(map) => {
                let name =
                    match map.get("name") {
                        Some(name) =>
                            match name {
                                &Value::String(ref name) => name,
                                _ => return Err(JsonError::invalid_type(Type::String)),
                            },
                        None => return Err(JsonError::missing_field("Missing \"name\"")),
                    };
                let id =
                    match map.get("id") {
                        Some(id) =>
                            match id {
                                &Value::I64(id) => id,
                                &Value::U64(id) => id as i64,
                                _ => return Err(JsonError::invalid_type(Type::I64)),
                            },
                        None => return Err(JsonError::missing_field("Missing \"id\"")),
                    };
                let address =
                    match map.get("address") {
                        Some(address) =>
                            match address {
                                &Value::Array(ref address) =>
                                    match address.len() {
                                        2 =>
                                            match (&address[0], &address[1]) {
                                                (&Value::String(ref laddress), &Value::String(ref raddress)) => (laddress.clone(), raddress.clone()),
                                                _ => return Err(JsonError::invalid_type(Type::String)),
                                            },
                                        _ => return Err(JsonError::invalid_length(2)),
                                    },
                                _ => return Err(JsonError::invalid_type(Type::String)),
                            },
                        None => return Err(JsonError::missing_field("Missing \"address\"")),
                    };

                Ok(UserInfo{
                    name: name.clone(),
                    id: id,
                    address: address,//(address.0.clone(), address.1.clone()),
                })
            },
            _ => Err(JsonError::invalid_type(Type::Struct)),
        }
    }
}

fn main() {
    let original = UserInfo{
        name: "user".to_owned(),
        id: 50030,
        address: ("Diagon Alley".to_owned(), "London, UK".to_owned()),
    };
    let serialized = original.to_json_string().unwrap();
    println!("{}", serialized);  // Space-efficient "ugly" format
    let deserialized = UserInfo::from_json_string(&serialized).unwrap();
    assert_eq!(original, deserialized);
}

Enums

JsonError
This type represents all possible errors that can occur when serializing or deserializing a value into JSON.

Traits

FromJsonnable
A trait for types supporting deserialization from JSON

ToJsonnable
A trait for types supporting serialization to JSON